I improved Add performance.
```
name         old time/op  new time/op  delta
AddSafe-8     155µs ± 1%   137µs ± 1%  -11.73%  (p=0.000 n=9+10)
AddUnsafe-8   145µs ± 3%   125µs ± 1%  -13.87%  (p=0.000 n=10+9)
```