@jeffwidman 2022-03-28
@orisano if these benchmarks are indeed still relevant against the new `generics`-based implementation, then this looks like a useful PR... 11%-13% improvement is nothing to sneeze at!